requirements.txt
flask>=2.x.y
google-genai>=0.x.y
google-generativeai
sentence-transformers>=2.x.y

faiss-cpu


